[ska] relocate platform banner, cloud, so, custom branding tests (#227269)
## Summary

Part of https://github.com/elastic/kibana-team/issues/1503

This PR is mostly about moving platform tests from x-pack/test

```
x-pack/test/
     | - banner_functional/
     | - cloud_integration/
     | - saved_objects_field_count/
     | - custom_branding/
```

After:
```
x-pack/platform/test/
     | - banner_functional/
     | - cloud_integration/
     | - saved_objects_field_count/
     | - custom_branding/
```

---------

(cherry picked from commit 0f5ca5ff10e8f0e282109b7692189066043b6e26)

# Conflicts:
#	.github/CODEOWNERS
#	x-pack/platform/test/functional/page_objects/index.ts
#	x-pack/platform/test/tsconfig.json
#	x-pack/test/functional/page_objects/index.ts
