+++ .buildkite/ftr_platform_stateful_configs.yml
@@ -32,7 +32,7 @@ disabled:
   - x-pack/platform/test/alerting_api_integration/security_and_spaces/group2/tests/actions/config.ts
   - x-pack/platform/test/alerting_api_integration/security_and_spaces/group2/tests/telemetry/config.ts
   - x-pack/platform/test/alerting_api_integration/spaces_only_legacy/config.ts
-  - x-pack/test/cloud_integration/config.ts
+  - x-pack/platform/test/cloud_integration/config.ts
   - x-pack/platform/test/load/config.ts
   - x-pack/platform/test/plugin_api_perf/config.ts
   - x-pack/test/screenshot_creation/config.ts
@@ -159,7 +159,7 @@ enabled:
   - x-pack/platform/test/alerting_api_integration/spaces_only/tests/action_task_params/config.ts
   - x-pack/test/api_integration/apis/cases/config.ts
   - x-pack/test/api_integration/apis/security/config.ts
-  - x-pack/test/banners_functional/config.ts
+  - x-pack/platform/test/banners_functional/config.ts
   - x-pack/platform/test/cases_api_integration/security_and_spaces/config_basic.ts
   - x-pack/platform/test/cases_api_integration/security_and_spaces/config_trial.ts
   - x-pack/platform/test/cases_api_integration/security_and_spaces/config_trial_common.ts
@@ -292,7 +292,7 @@ enabled:
   - x-pack/test/saved_object_tagging/api_integration/tagging_api/config.ts
   - x-pack/test/saved_object_tagging/api_integration/tagging_usage_collection/config.ts
   - x-pack/test/saved_object_tagging/functional/config.ts
-  - x-pack/test/saved_objects_field_count/config.ts
+  - x-pack/platform/test/saved_objects_field_count/config.ts
   - x-pack/platform/test/search_sessions_integration/config.ts
   - x-pack/test/security_api_integration/anonymous_es_anonymous.config.ts
   - x-pack/test/security_api_integration/anonymous.config.ts
@@ -361,7 +361,7 @@ enabled:
   - x-pack/performance/journeys_e2e/cloud_security_dashboard.ts
   - x-pack/performance/journeys_e2e/apm_service_inventory.ts
   - x-pack/performance/journeys_e2e/infra_hosts_view.ts
-  - x-pack/test/custom_branding/config.ts
+  - x-pack/platform/test/custom_branding/config.ts
   # configs migrated to the new Kibana architecture
   - x-pack/platform/test/api_integration/apis/aiops/config.ts
   - x-pack/platform/test/api_integration/apis/cloud/config.ts
