+++ .buildkite/scripts/saved_object_field_metrics.sh
@@ -8,4 +8,4 @@ echo '--- Default Saved Object Field Metrics'
 node scripts/functional_tests \
   --debug --bail \
   --kibana-install-dir "$KIBANA_BUILD_LOCATION" \
-  --config x-pack/test/saved_objects_field_count/config.ts
+  --config x-pack/platform/test/saved_objects_field_count/config.ts
