+++ package.json
@@ -232,7 +232,7 @@
     "@kbn/cloud-data-migration-plugin": "link:x-pack/platform/plugins/private/cloud_integrations/cloud_data_migration",
     "@kbn/cloud-experiments-plugin": "link:x-pack/platform/plugins/private/cloud_integrations/cloud_experiments",
     "@kbn/cloud-full-story-plugin": "link:x-pack/platform/plugins/private/cloud_integrations/cloud_full_story",
-    "@kbn/cloud-integration-saml-provider-plugin": "link:x-pack/test/cloud_integration/plugins/saml_provider",
+    "@kbn/cloud-integration-saml-provider-plugin": "link:x-pack/platform/test/cloud_integration/plugins/saml_provider",
     "@kbn/cloud-links-plugin": "link:x-pack/platform/plugins/private/cloud_integrations/cloud_links",
     "@kbn/cloud-plugin": "link:x-pack/platform/plugins/shared/cloud",
     "@kbn/cloud-security-posture": "link:x-pack/solutions/security/packages/kbn-cloud-security-posture/public",
