+++ tsconfig.base.json
@@ -184,8 +184,8 @@
       "@kbn/cloud-experiments-plugin/*": ["x-pack/platform/plugins/private/cloud_integrations/cloud_experiments/*"],
       "@kbn/cloud-full-story-plugin": ["x-pack/platform/plugins/private/cloud_integrations/cloud_full_story"],
       "@kbn/cloud-full-story-plugin/*": ["x-pack/platform/plugins/private/cloud_integrations/cloud_full_story/*"],
-      "@kbn/cloud-integration-saml-provider-plugin": ["x-pack/test/cloud_integration/plugins/saml_provider"],
-      "@kbn/cloud-integration-saml-provider-plugin/*": ["x-pack/test/cloud_integration/plugins/saml_provider/*"],
+      "@kbn/cloud-integration-saml-provider-plugin": ["x-pack/platform/test/cloud_integration/plugins/saml_provider"],
+      "@kbn/cloud-integration-saml-provider-plugin/*": ["x-pack/platform/test/cloud_integration/plugins/saml_provider/*"],
       "@kbn/cloud-links-plugin": ["x-pack/platform/plugins/private/cloud_integrations/cloud_links"],
       "@kbn/cloud-links-plugin/*": ["x-pack/platform/plugins/private/cloud_integrations/cloud_links/*"],
       "@kbn/cloud-plugin": ["x-pack/platform/plugins/shared/cloud"],
