+++ x-pack/platform/test/banners_functional/ftr_provider_context.ts
