+++ x-pack/platform/test/banners_functional/tests/global.ts
