+++ x-pack/platform/test/banners_functional/tests/spaces.ts
