+++ x-pack/platform/test/cloud_integration/config.ts
