+++ x-pack/platform/test/cloud_integration/ftr_provider_context.d.ts
