+++ x-pack/platform/test/cloud_integration/plugins/saml_provider/kibana.jsonc
