+++ x-pack/platform/test/cloud_integration/plugins/saml_provider/server/saml_tools.ts
