+++ x-pack/platform/test/cloud_integration/plugins/saml_provider/tsconfig.json
@@ -1,5 +1,5 @@
 {
-  "extends": "../../../../../tsconfig.base.json",
+  "extends": "../../../../../../tsconfig.base.json",
   "compilerOptions": {
     "outDir": "target/types"
   },
