+++ x-pack/platform/test/cloud_integration/tests/fullstory.ts
