+++ x-pack/platform/test/custom_branding/config.ts
@@ -11,7 +11,7 @@ import { services, pageObjects } from './ftr_provider_context';
 
 export default async function ({ readConfigFile }: FtrConfigProviderContext) {
   const kibanaFunctionalConfig = await readConfigFile(
-    require.resolve('../functional/config.base.js')
+    require.resolve('../functional/config.base.ts')
   );
 
   return {
@@ -24,21 +24,21 @@ export default async function ({ readConfigFile }: FtrConfigProviderContext) {
     pageObjects,
 
     junit: {
-      reportName: 'X-Pack Banners Functional Tests',
+      reportName: 'X-Pack Custom Branding Functional Tests',
     },
 
-    esTestCluster: kibanaFunctionalConfig.get('esTestCluster'),
+    esTestCluster: {
+      ...kibanaFunctionalConfig.get('esTestCluster'),
+      license: 'trial',
+      serverArgs: [`xpack.license.self_generated.type='trial'`],
+    },
     apps: {
       ...kibanaFunctionalConfig.get('apps'),
     },
 
     kbnTestServer: {
       ...kibanaFunctionalConfig.get('kbnTestServer'),
-      serverArgs: [
-        ...kibanaFunctionalConfig.get('kbnTestServer.serverArgs'),
-        '--xpack.banners.placement=top',
-        '--xpack.banners.textContent="global banner text"',
-      ],
+      serverArgs: [...kibanaFunctionalConfig.get('kbnTestServer.serverArgs')],
     },
   };
 }
