+++ x-pack/platform/test/custom_branding/ftr_provider_context.ts
