+++ x-pack/platform/test/custom_branding/tests/acme_text.png
