+++ x-pack/platform/test/custom_branding/tests/index.ts
