+++ x-pack/platform/test/custom_branding/tests/settings.ts
