+++ x-pack/platform/test/functional/page_objects/banners_page.ts
