+++ x-pack/platform/test/tsconfig.json
@@ -127,5 +127,14 @@
     "@kbn/core-provider-plugin",
     "@kbn/reporting-plugin",
     "@kbn/kibana-usage-collection-plugin",
+    "@kbn/aiops-change-point-detection",
+    "@kbn/aiops-plugin",
+    "@kbn/ml-category-validator",
+    "@kbn/response-ops-alerts-filters-form",
+    "@kbn/embeddable-alerts-table-plugin",
+    "@kbn/apm-synthtrace-client",
+    "@kbn/apm-synthtrace",
+    "@kbn/cloud-integration-saml-provider-plugin",
+    "@kbn/ci-stats-reporter",
   ]
 }
