+++ x-pack/test/functional/page_objects/index.ts
@@ -23,9 +23,10 @@ import { SnapshotRestorePageProvider } from '@kbn/test-suites-xpack-platform/fun
 import { UserProfilePageProvider } from '@kbn/test-suites-xpack-platform/functional/page_objects/user_profile_page';
 import { SearchSessionsPageProvider } from '@kbn/test-suites-xpack-platform/functional/page_objects/search_sessions_management_page';
 import { GraphPageObject } from '@kbn/test-suites-xpack-platform/functional/page_objects/graph_page';
+import { MaintenanceWindowsPageProvider } from '@kbn/test-suites-xpack-platform/functional/page_objects/maintenance_windows_page';
+import { BannersPageObject } from '@kbn/test-suites-xpack-platform/functional/page_objects/banners_page';
 import { ApiKeysPageProvider } from './api_keys_page';
 import { AssetDetailsProvider } from './asset_details';
-import { BannersPageObject } from './banners_page';
 import { CrossClusterReplicationPageProvider } from './cross_cluster_replication_page';
 import { DetectionsPageObject } from '../../security_solution_ftr/page_objects/detections';
 import { EmbeddedConsoleProvider } from './embedded_console';
@@ -41,7 +42,6 @@ import { InfraSavedViewsProvider } from './infra_saved_views';
 import { IngestPipelinesPageProvider } from './ingest_pipelines_page';
 import { LicenseManagementPageProvider } from './license_management_page';
 import { LogstashPageObject } from './logstash_page';
-import { MaintenanceWindowsPageProvider } from './maintenance_windows_page';
 import { NavigationalSearchPageObject } from './navigational_search';
 import { ObservabilityLogsExplorerPageObject } from './observability_logs_explorer';
 import { DatasetQualityPageObject } from './dataset_quality';
