+++ x-pack/test/tsconfig.json
@@ -55,7 +55,6 @@
     "@kbn/dev-utils",
     "@kbn/ftr-common-functional-services",
     "@kbn/expect",
-    "@kbn/ci-stats-reporter",
     "@kbn/std",
     "@kbn/apm-synthtrace",
     "@kbn/core-saved-objects-base-server-internal",
@@ -81,7 +80,6 @@
     "@kbn/core-saved-objects-server",
     "@kbn/user-profile-components",
     "@kbn/apm-synthtrace-client",
-    "@kbn/cloud-integration-saml-provider-plugin",
     "@kbn/security-api-integration-helpers",
     "@kbn/alerts-as-data-utils",
     "@kbn/discover-plugin",
